backend/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
gitingest
pydantic